backend/requirements.txt


flask
pynacl
